rename just about everything to deathshamer
