+++ README.md
@@ -1,15 +1,17 @@
 # Death Shamer
-Takes a screenshot of deaths. Also supports discord webhook integration.
+Takes a screenshot of deaths. Also supports discord webhook integration. Previously named Raid Shamer.
 
 You can toggle whether this shames your own deaths.
 You can toggle whether to shame all deaths (default) or only your friends' deaths.
-You can toggle whether this shames people on a per-raid basis or outside of raids.
+You can toggle whether this shames people on a per-raid basis and/or outside of raids.
 
 This plugin by default just saves a photo of a death to a folder called "Wall of Shame." I'm assuming if you're reading this, you are looking for discord webhook information.
 
 ## Example Discord Webhook message
 
-![Discord webook example](https://i.imgur.com/7mlQ2Pi.png)
+Note: This particular screenshot is compressed/low quality due to editing (adding the censors). Screenshots taken by the plugin are crisp and uncensored.
+
+![Discord webook example](https://i.imgur.com/l83BtsS.png)
 
 ## How to set up the webhook
 
@@ -19,10 +21,10 @@ Click the settings gear beside a channel, then integrations, and then "View Webh
 
 Enter in a name, select the channel, then optionally upload a profile photo. Save your changes and click "Copy Webhook URL."
 
-![Discord webhook setup](https://i.imgur.com/Y5kTDU4.png)
+![Discord webhook setup](https://i.imgur.com/cWRSh8v.png)
 
 Open up Runelite and go to the settings for Death Shamer. Click the box to turn on Discord Webhook support and then paste the URL into the text box.
 
-![Runelite webhook settings](https://i.imgur.com/v1XOjiM.png)
+![Runelite webhook settings](https://i.imgur.com/YYUZGp3.png)
 
 This is not malicious and your webhook will never be sent to anyone else. That's all there is to it.
