+++ build.gradle
@@ -25,7 +25,7 @@ dependencies {
 	}
 }
 
-group = 'ejedev.raidshamer'
+group = 'ejedev.deathshamer'
 version = '1.0'
 sourceCompatibility = '1.8'
