+++ runelite-plugin.properties
@@ -1,6 +1,6 @@
 displayName=Death Shamer
 author=Botanophobia
 support=https://github.com/jack0lantern/raidshamer
-description=Takes a screenshot of deaths during bosses and raids. Supports discord webhook integration.
+description=Takes a screenshot of deaths. Supports discord webhook integration.
 tags=death, raid, raids, shame, tob, theater, cox, chambers, discord, webhook
-plugins=ejedev.raidshamer.RaidShamerPlugin
+plugins=ejedev.deathshamer.DeathShamerPlugin
