+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'raid-shamer'
+rootProject.name = 'death-shamer'
